Moves description update form and logic to a separate component. Removes unused variables, elements and functions from HostProfile component.
